+++ tests/branches/valve/test_orange_box.py
@@ -9,9 +9,11 @@ bsps = utils.get_test_maps(ValveBsp, {orange_box: ["Team Fortress 2"]})
 
 
 class TestMethods:
+    displacement_bsps = [b for b in bsps if b.headers["DISPLACEMENT_INFO"].length != 0]
+
     # TODO: def test_vertices_of_face(bsp: ValveBsp):
 
-    @pytest.mark.parametrize("bsp", bsps, ids=[b.filename for b in bsps])
+    @pytest.mark.parametrize("bsp", displacement_bsps, ids=[b.filename for b in displacement_bsps])
     def test_vertices_of_displacement(self, bsp: ValveBsp):
         for disp_info in getattr(bsp, "DISPLACEMENT_INFO", list()):
             face_index = disp_info.face