+++ tests/test_BspLump.py
@@ -1,15 +1,16 @@
-import pytest
+from . import utils
 
-from bsp_tool import load_bsp, lumps
+from bsp_tool import lumps
+from bsp_tool import ValveBsp, IdTechBsp
 from bsp_tool.branches.id_software import quake, quake3
+from bsp_tool.branches.valve import orange_box
+
+import pytest
 
 # TODO: collect valid lumps of desired type for each test, rather than hardcoded lump names
 
-global bsps
-bsps = {"q3_lobby": load_bsp("tests/maps/Quake 3 Arena/mp_lobby.bsp"),
-        "tf2_test2": load_bsp("tests/maps/Team Fortress 2/test2.bsp"),
-        "tf2_test_displacement_decompile": load_bsp("tests/maps/Team Fortress 2/test_displacement_decompile.bsp"),
-        "tf2_test_physcollide": load_bsp("tests/maps/Team Fortress 2/test_physcollide.bsp")}
+bsps = [*utils.get_test_maps(ValveBsp, {orange_box: ["Team Fortress 2"]}),
+        *utils.get_test_maps(IdTechBsp, {quake3: ["Quake 3 Arena"]})]
 
 
 def raw_lump_of(bsp) -> lumps.RawBspLump:
@@ -22,72 +23,66 @@ def raw_lump_of(bsp) -> lumps.RawBspLump:
 
 
 class TestRawBspLump:
-    raw_lumps = list(map(raw_lump_of, bsps.values()))
+    raw_lumps = list(map(raw_lump_of, bsps))
 
-    def test_its_raw(self):
-        for lump in self.raw_lumps:
-            assert isinstance(lump, lumps.RawBspLump), f"it's not raw! it's {type(lump)}"
+    @pytest.mark.parametrize("raw_lump", raw_lumps, ids=[b.filename for b in bsps])
+    def test_its_raw(self, raw_lump):
+        assert isinstance(raw_lump, lumps.RawBspLump)
 
-    def test_list_conversion(self):
-        for map_name, lump in zip(bsps, self.raw_lumps):
-            assert list(lump) == [int(b) for b in lump], f"{map_name} failed"
+    @pytest.mark.parametrize("raw_lump", raw_lumps, ids=[b.filename for b in bsps])
+    def test_list_conversion(self, raw_lump):
+        assert list(raw_lump) == [int(b) for b in raw_lump]
 
-    def test_indexing(self):
-        for map_name, lump in zip(bsps, self.raw_lumps):
-            assert isinstance(lump[0], int), f"{map_name} failed"
-            assert isinstance(lump[:1], bytes), f"{map_name} failed"
-            assert len(lump[:1]) == 1, f"{map_name} failed"
-            # ^ all three just look at the first byte
-            # expecting behaviour the same as if lump was a bytestring
-            assert len(lump[-2:]) == 2, f"{map_name} failed"
-            with pytest.raises(TypeError):
-                assert lump["one"], f"{map_name} failed"
+    @pytest.mark.parametrize("raw_lump", raw_lumps, ids=[b.filename for b in bsps])
+    def test_indexing(self, raw_lump):
+        assert isinstance(raw_lump[0], int)
+        assert isinstance(raw_lump[:1], bytearray)
+        assert len(raw_lump[:1]) == 1
+        assert len(raw_lump[-2:]) == 2
+        with pytest.raises(TypeError):
+            assert raw_lump["one"]
 
 
 class TestBspLump:
-    def test_list_conversion(self):
-        for map_name in bsps:
-            lump = bsps[map_name].VERTICES
-            assert list(lump) == [b for b in lump], f"{map_name}.VERTICES failed"
+    @pytest.mark.parametrize("bsp", bsps, ids=[b.filename for b in bsps])
+    def test_list_conversion(self, bsp):
+        lump = bsp.VERTICES
+        assert list(lump) == [b for b in lump]
 
-    def test_indexing(self):
-        for map_name in bsps:
-            lump = bsps[map_name].VERTICES
-            LumpClass = quake.Vertex if map_name != "q3_lobby" else quake3.Vertex
-            assert isinstance(lump[0], LumpClass), f"{map_name} failed"
-            assert isinstance(lump[:1], list), f"{map_name} failed"
-            assert len(lump[:1]) == 1, f"{map_name} failed"
-            # ^ all three just look at the first byte
-            # expecting behaviour the same as if lump was a bytestring
-            assert len(lump[-2:]) == 2, f"{map_name} failed"
-            # TODO: check negative indices line up [_remap_negative_index]
-            # TODO: check slice cases (negative step, wide step, invalid slice) [_remap_slice]
-            with pytest.raises(TypeError):
-                assert lump["one"], f"{map_name} failed"
+    @pytest.mark.parametrize("bsp", bsps, ids=[b.filename for b in bsps])
+    def test_indexing(self, bsp):
+        lump = bsp.VERTICES
+        LumpClass = quake.Vertex if bsp.branch != quake3 else quake3.Vertex
+        assert isinstance(lump[0], LumpClass)
+        assert isinstance(lump[:1], list)
+        assert len(lump[:1]) == 1
+        assert len(lump[-2:]) == 2
+        with pytest.raises(TypeError):
+            assert lump["one"]
 
-    def test_del(self):
-        for map_name in bsps:
-            lump = bsps[map_name].VERTICES
-            initial_length = len(lump)
-            del lump[0]
-            assert len(lump) == initial_length - 1, f"{map_name} failed"
-            initial_length = len(lump)
-            del lump[:2]
-            assert len(lump) == initial_length - 2, f"{map_name} failed"
+    @pytest.mark.parametrize("bsp", bsps, ids=[b.filename for b in bsps])
+    def test_del(self, bsp):
+        lump = bsp.VERTICES
+        initial_length = len(lump)
+        # delete single index
+        del lump[0]
+        assert len(lump) == initial_length - 1
+        # delete slice
+        initial_length = len(lump)
+        del lump[:2]
+        assert len(lump) == initial_length - 2
 
-    def test_setitem(self):
-        for map_name in bsps:
-            lump = bsps[map_name].VERTICES
-            empty_entry = lump.LumpClass()
-            lump[0] = empty_entry
-            assert lump[0] == empty_entry, f"{map_name} failed"
-            lump[:2] = [empty_entry, empty_entry]
-            assert lump[:2] == [empty_entry, empty_entry], f"{map_name} failed"
-            # TODO: allow for insert via slice & test for this
-            # TODO: test changes to object attrs for Issue #23
-            # -- e.g. bsp.LUMP[index].attr = val (uses soft copies)
-
-    # TODO: test_iadd (__iadd__ method; overrides +=)
+    @pytest.mark.parametrize("bsp", bsps, ids=[b.filename for b in bsps])
+    def test_setitem(self, bsp):
+        lump = bsp.VERTICES
+        empty_entry = lump.LumpClass()
+        lump[0] = empty_entry
+        assert lump[0] == empty_entry
+        lump[:2] = [empty_entry, empty_entry]
+        assert lump[:2] == [empty_entry, empty_entry]
+        # TODO: allow for insert via slice & test for this
+        # TODO: test changes to object attrs for Issue #23
+        # -- e.g. bsp.LUMP[index].attr = val (uses soft copies)
 
 
 class TestExternalBspLump:  # TODO: ship bespoke RespawnBsp .bsp_lump files with tests
@@ -95,29 +90,26 @@ class TestExternalBspLump:  # TODO: ship bespoke RespawnBsp .bsp_lump files with
 
 
 class TestBasicBspLump:
-    def test_its_basic(self):
-        for map_name in bsps:
-            lump = bsps[map_name].LEAF_FACES
-            assert isinstance(lump, lumps.BasicBspLump), f"it's not basic! it's {type(lump)}"
+    @pytest.mark.parametrize("bsp", bsps, ids=[b.filename for b in bsps])
+    def test_its_basic(self, bsp):
+        lump = bsp.LEAF_FACES
+        assert isinstance(lump, lumps.BasicBspLump), type(lump)
 
-    def test_list_conversion(self):
-        for map_name in bsps:
-            lump = bsps[map_name].LEAF_FACES
-            assert list(lump) == [b for b in lump], f"{map_name} vis lump does not match"
+    @pytest.mark.parametrize("bsp", bsps, ids=[b.filename for b in bsps])
+    def test_list_conversion(self, bsp):
+        lump = bsp.LEAF_FACES
+        assert list(lump) == [b for b in lump]
 
-    def test_indexing(self):
+    @pytest.mark.parametrize("bsp", bsps, ids=[b.filename for b in bsps])
+    def test_indexing(self, bsp):
         for map_name in bsps:
-            lump = bsps[map_name].LEAF_FACES
-            LumpClass = bsps[map_name].branch.BASIC_LUMP_CLASSES["LEAF_FACES"]
+            lump = bsp.LEAF_FACES
+            LumpClass = bsp.branch.BASIC_LUMP_CLASSES["LEAF_FACES"]
             if isinstance(LumpClass, dict):  # ValveBsp branches use dicts for multiple lump versions
-                LumpClass = list(LumpClass.values())[0]
-            assert isinstance(lump[0], LumpClass), f"{map_name} failed"
-            assert isinstance(lump[:1], list), f"{map_name} failed"
-            assert len(lump[:1]) == 1, f"{map_name} failed"
-            # ^ all three just look at the first byte
-            # expecting behaviour the same as if lump was a bytestring
-            assert len(lump[-2:]) == 2, f"{map_name} failed"
-            # TODO: check negative indices line up
-            # TODO: check slice cases (negative step, wide step, invalid slice)
+                LumpClass = LumpClass[bsp.headers["LEAF_FACES"].version]
+            assert isinstance(lump[0], LumpClass)
+            assert isinstance(lump[:1], list)
+            assert len(lump[:1]) == 1
+            assert len(lump[-2:]) == 2
             with pytest.raises(TypeError):
-                assert lump["one"], f"{map_name} failed"
+                assert lump["one"]